avoid small files (merge content with existing files)
